# cmd_default/http/api_docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "api example",
        "title": "api example",
        "termsOfService": "https://domain.com",
        "contact": {
            "name": "Support",
            "url": "https://domain.com",
            "email": "contact@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/kv/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "creat key value pair",
                "parameters": [
                    {
                        "description": "creat key value pair",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Msg_Req_CreateKv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result",
                        "schema": {
                            "$ref": "#/definitions/api.API_META_STATUS"
                        }
                    }
                }
            }
        },
        "/api/kv/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "delete key value pair",
                "parameters": [
                    {
                        "description": "delete key value pair",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Msg_Req_DeleteKv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result",
                        "schema": {
                            "$ref": "#/definitions/api.API_META_STATUS"
                        }
                    }
                }
            }
        },
        "/api/kv/query": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "query key value pair",
                "parameters": [
                    {
                        "description": "query key value pair",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Msg_Req_QueryKv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result",
                        "schema": {
                            "$ref": "#/definitions/api.Msg_Resp_QueryKv"
                        }
                    }
                }
            }
        },
        "/api/kv/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "update key value pair",
                "parameters": [
                    {
                        "description": "update key value pair",
                        "name": "msg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Msg_Req_UpdateKv"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result",
                        "schema": {
                            "$ref": "#/definitions/api.API_META_STATUS"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.API_META_STATUS": {
            "type": "object",
            "properties": {
                "meta_message": {
                    "type": "string"
                },
                "meta_status": {
                    "type": "integer"
                }
            }
        },
        "api.DBKV": {
            "type": "object",
            "properties": {
                "created_unixtime": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "update_unixtime": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "api.Msg_Req_CreateKv": {
            "description": "Msg_Req_CreateRule",
            "type": "object",
            "properties": {
                "description": {
                    "description": "required",
                    "type": "string"
                },
                "key": {
                    "description": "required",
                    "type": "string"
                },
                "value": {
                    "description": "required",
                    "type": "string"
                }
            }
        },
...
        "api.Msg_Req_DeleteKv": {
            "description": "Msg_Req_DeleteKv",
            "type": "object",
            "properties": {
                "filter": {
                    "description": "required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Msg_Req_DeleteKv_Filter"
                        }
                    ]
                }
            }
        },
        "api.Msg_Req_DeleteKv_Filter": {
            "description": "Msg_Req_DeleteKv_Filter",
            "type": "object",
            "properties": {
                "keys": {
                    "description": "required",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.Msg_Req_QueryKv": {
            "description": "Msg_Req_QueryKv",
            "type": "object",
            "properties": {
                "filter": {
                    "description": "required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Msg_Req_QueryKv_Filter"
                        }
                    ]
                }
            }
        },
        "api.Msg_Req_QueryKv_Filter": {
            "description": "Msg_Req_QueryKv_Filter",
            "type": "object",
            "properties": {
                "keys": {
                    "description": "optional",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.Msg_Req_UpdateKv": {
            "description": "Msg_Req_UpdateKv",
            "type": "object",
            "properties": {
                "filter": {
                    "description": "required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Msg_Req_UpdateKv_Filter"
                        }
                    ]
                },
                "update": {
                    "description": "required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Msg_Req_UpdateKv_To"
                        }
                    ]
                }
            }
        },
        "api.Msg_Req_UpdateKv_Filter": {
            "description": "Msg_Req_UpdateKv_Filter",
            "type": "object",
            "properties": {
                "key": {
                    "description": "required",
                    "type": "string"
                }
            }
        },
        "api.Msg_Req_UpdateKv_To": {
            "description": "Msg_Req_UpdateKv_To",
            "type": "object",
            "properties": {
                "description": {
                    "description": "required",
                    "type": "string"
                },
                "value": {
                    "description": "required",
                    "type": "string"
                }
            }
        },
...
        "api.Msg_Resp_QueryKv": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.DBKV"
                    }
                },
                "meta_message": {
                    "type": "string"
                },
                "meta_status": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}